Display table with averages and other modifications
Using pandas to display a table with the time-averaged values.

Strouhal number is averaged over a given number of last oscillations
of the lift curve that ends at a given time.

Change floag to add a simulation to compare with (`--add-simulation`
instead of `--others`). To add two simulations, use the flag twice.

Minor modifications through the code. (Should not impact the user.)