Add argument to specify the forces folder
With the argument `forces_folder`, you can specify the relative path of the folder
containing the forces.
The default value is 'postProcessing/forces'.
Same for the force coefficients with the argument `force_coefficients_folder`.
Its default value is 'postProcessing/forceCoeffs'.

This modification facilitates the processing of simulations run with different OpenFOAM versions.